// src/app/models/mrp/mrp-dashboard-widget-type.enum.ts
export enum MrpDashboardWidgetType {
    PRODUCTION_PLAN_FULFILLMENT_BY_COMPANY = 27,
    PRODUCTION_PLAN_FULFILLMENT_DETAIL_BY_COMPANY = 28,
    PRODUCTION_PLAN_FULFILLMENT_BY_BRANCHOFFICE = 29,
    PRODUCTION_PLAN_FULFILLMENT_DETAIL_BY_BRANCHOFFICE = 30,
    PRODUCTION_ORDER_PROCESSED_BY_ROOM_COMPANY = 61,
    PRODUCTION_ORDER_PROCESSED_BY_ROOM_BRANCHOFFICE = 62,
    ROOM_EFFICIENCY_BY_COMPANY = 63,
    ROOM_EFFICIENCY_BY_BRANCHOFFICE = 64,
    PRODUCTION_ORDER_BY_STATUS_COMPANY = 65,
    PRODCUTION_ORDER_BY_STATUS_BRANCHOFFICE = 66,
    COST_CHANGED_MRP_MPC_BY_COMPANY = 112,
    COST_CHANGED_MRP_MPC_BY_BRANCHOFFICE = 113,
    DERIVATE_ROOM_PERFORMANCE_BY_COMPANY = 129,
    DERIVATE_ROOM_PERFORMANCE_BY_BRANCHOFFICE = 130
}